requirements.txt
langchain==0.2.0
langchain-ollama==0.1.0
langchain-chroma==0.1.1
pypdf==4.1.0
sentence-transformers==2.2.2
python-docx==0.8.11
python-pptx==0.6.21
openpyxl==3.1.2
pandas==2.2.1
PyYAML==6.0.1
matplotlib==3.8.3